Add second memory test using fixed length keys
Second memory test is in separate file, factored out common setup
code.

New test uses a fixed length key distribution, with most unique values
in the first digit. These keys are not amenable to prefix compression,
and take up more space than a key distribution with most unique values
in the last digit. See
https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=269919fa7c85bc6476accdcca2887298
for an example.
